+++ src/errors.ts
@@ -6,7 +6,12 @@ export class GenericError extends Error {
   }
 }
 export class AuthenticationError extends GenericError {
-  constructor(error: string, error_description: string, public state: string) {
+  constructor(
+    error: string,
+    error_description: string,
+    public state: string,
+    public appState: any = null
+  ) {
     super(error, error_description);
     //https://github.com/Microsoft/TypeScript-wiki/blob/master/Breaking-Changes.md#extending-built-ins-like-error-array-and-map-may-no-longer-work
     Object.setPrototypeOf(this, AuthenticationError.prototype);