+++ src/global.ts
@@ -269,6 +269,7 @@ export interface TokenEndpointOptions {
 export interface OAuthTokenOptions extends TokenEndpointOptions {
   code_verifier: string;
   code: string;
+  redirect_uri: string;
 }
 
 /**